requirements.txt
Flask>=3.0
supabase>=2.0
python-dotenv>=1.0
gunicorn>=22.0
requests>=2.31
beautifulsoup4>=4.12
